Add "smart caching" to irods-build image
Allows irods-build-and-test-workflow to selectively use cached irods-build
images based on git commit resolution. This means that branch names can be
used for the commitish field when running tests and the image will rebuild
or not depending on if the commit's SHA at the branch head changed.

NOTE: This change will break your Jenkins instance unless you install the
pygithub pip package in your running instance or rebuild the Jenkins image
and re-push to the registry.