[#79][#82] Fix enable_ssl in topology tests
Sleeps have been inserted around logic for enabling SSL on the iRODS
servers in a topology. A non-timing based solution should be implemented
but this is a placeholder to allow for SSL testing to proceed within the
Consortium.

Prevents generating SSL files when enable_ssl is false. Also prevents SSL
files from being generated and copied within each docker container in a
topology. Now the files are generated once inside the catalog provider
and copied to the catalog consumers.