[29] Parameterize mysql odbc connector files
This change removes the volume mount for the file in RENCI's
/projects NFS mount. Now CI will pass the MySQL ODBC Connector
file locations as parameters which are then used by the scripts
to setup the MySQL client.

To change the default behavior, change the parameter for the job
being run called PARAMETER_MYSQL_ODBC_CONNECTOR_DIR to a location
accessible for read by the host machine. This will become a mount
point inside of the containers running tests. The location must have
a file corresponding to the filenames in the map in configuration.py
called mysql_odbc_connectors (there is one entry per supported platform).
If the file is called something different, or a different version is
desired, the filename string in the map should be changed to the
desired value.