[#130] Get plugin tests running.
Also: To avoid introducing any extra indentation that could
conflict with a revert of the recent 463d (container removal)
for debugging, we use an RAII not depending on try/finally to
get the command output in the "run_command_in_container" function
to be printed to the Jenkins console.  This "alternative way" just
relies on the __del__ method of an appropriately scoped object to
be automatically called when the run_command_in_container function
exits.