directory and project name checks
Measures are taken now to ensure local repository is an absolute
directory and that hints to the docker-compose project (to be run as
the test) point to a location subordinate to that directory.
The project name must also be unique, so for that we key on the
top level directory name, which itself is ensured to be unique.

Furthermore, if the indicated local repository name already exists,
we make a new one with an affixed, base32-encoded timestamp.